[mob] Move toast
