+++ mobile/lib/services/account/user_service.dart
@@ -40,10 +40,10 @@ import 'package:photos/ui/account/two_factor_authentication_page.dart';
 import 'package:photos/ui/account/two_factor_recovery_page.dart';
 import 'package:photos/ui/account/two_factor_setup_page.dart';
 import "package:photos/ui/common/progress_dialog.dart";
+import 'package:photos/ui/notification/toast.dart';
 import "package:photos/ui/tabs/home_widget.dart";
 import 'package:photos/utils/dialog_util.dart';
 import 'package:photos/utils/navigation_util.dart';
-import 'package:photos/utils/toast_util.dart';
 import "package:pointycastle/export.dart";
 import "package:pointycastle/srp/srp6_client.dart";
 import "package:pointycastle/srp/srp6_standard_groups.dart";
