+++ mobile/lib/ui/account/delete_account_page.dart
@@ -12,8 +12,8 @@ import 'package:photos/services/account/user_service.dart';
 import 'package:photos/theme/ente_theme.dart';
 import 'package:photos/ui/components/buttons/button_widget.dart';
 import 'package:photos/ui/components/models/button_type.dart';
+import "package:photos/ui/notification/toast.dart";
 import 'package:photos/utils/dialog_util.dart';
-import "package:photos/utils/toast_util.dart";
 
 class DeleteAccountPage extends StatefulWidget {
   const DeleteAccountPage({
