+++ mobile/lib/ui/account/passkey_page.dart
@@ -12,9 +12,9 @@ import 'package:photos/services/account/user_service.dart';
 import "package:photos/ui/account/two_factor_authentication_page.dart";
 import "package:photos/ui/components/buttons/button_widget.dart";
 import "package:photos/ui/components/models/button_type.dart";
+import "package:photos/ui/notification/toast.dart";
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/navigation_util.dart";
-import "package:photos/utils/toast_util.dart";
 import "package:uni_links/uni_links.dart";
 import 'package:url_launcher/url_launcher_string.dart';
