+++ mobile/lib/ui/account/password_entry_page.dart
@@ -17,10 +17,10 @@ import 'package:photos/ui/account/recovery_key_page.dart';
 import 'package:photos/ui/common/dynamic_fab.dart';
 import 'package:photos/ui/common/web_page.dart';
 import "package:photos/ui/components/models/button_type.dart";
+import 'package:photos/ui/notification/toast.dart';
 import 'package:photos/ui/payment/subscription.dart';
 import 'package:photos/utils/dialog_util.dart';
 import 'package:photos/utils/navigation_util.dart';
-import 'package:photos/utils/toast_util.dart';
 import "package:styled_text/styled_text.dart";
 
 enum PasswordEntryMode {
