+++ mobile/lib/ui/account/sessions_page.dart
@@ -7,9 +7,9 @@ import 'package:photos/models/api/user/sessions.dart';
 import 'package:photos/services/account/user_service.dart';
 import "package:photos/theme/ente_theme.dart";
 import 'package:photos/ui/common/loading_widget.dart';
+import 'package:photos/ui/notification/toast.dart';
 import "package:photos/utils/date_time_util.dart";
 import 'package:photos/utils/dialog_util.dart';
-import 'package:photos/utils/toast_util.dart';
 
 class SessionsPage extends StatefulWidget {
   const SessionsPage({super.key});
