+++ mobile/lib/ui/account/two_factor_setup_page.dart
@@ -9,8 +9,8 @@ import "package:photos/generated/l10n.dart";
 import 'package:photos/services/account/user_service.dart';
 import 'package:photos/ui/account/recovery_key_page.dart';
 import 'package:photos/ui/lifecycle_event_handler.dart';
+import 'package:photos/ui/notification/toast.dart';
 import 'package:photos/utils/navigation_util.dart';
-import 'package:photos/utils/toast_util.dart';
 import "package:pinput/pinput.dart";
 
 class TwoFactorSetupPage extends StatefulWidget {
