+++ mobile/lib/ui/settings/debug/debug_section_widget.dart
@@ -9,8 +9,8 @@ import 'package:photos/theme/ente_theme.dart';
 import 'package:photos/ui/components/captioned_text_widget.dart';
 import 'package:photos/ui/components/expandable_menu_item_widget.dart';
 import 'package:photos/ui/components/menu_item_widget/menu_item_widget.dart';
+import 'package:photos/ui/notification/toast.dart';
 import 'package:photos/ui/settings/common_settings.dart';
-import 'package:photos/utils/toast_util.dart';
 
 class DebugSectionWidget extends StatelessWidget {
   const DebugSectionWidget({super.key});
