+++ mobile/lib/ui/sharing/add_participant_page.dart
@@ -15,9 +15,9 @@ import 'package:photos/ui/components/menu_item_widget/menu_item_widget.dart';
 import 'package:photos/ui/components/menu_section_description_widget.dart';
 import 'package:photos/ui/components/menu_section_title.dart';
 import 'package:photos/ui/components/models/button_type.dart';
+import "package:photos/ui/notification/toast.dart";
 import 'package:photos/ui/sharing/user_avator_widget.dart';
 import "package:photos/ui/sharing/verify_identity_dialog.dart";
-import "package:photos/utils/toast_util.dart";
 
 class AddParticipantPage extends StatefulWidget {
   final Collection collection;
