+++ mobile/lib/ui/viewer/file/video_widget_media_kit_new.dart
@@ -16,11 +16,11 @@ import "package:photos/services/files_service.dart";
 import "package:photos/theme/colors.dart";
 import "package:photos/ui/actions/file/file_actions.dart";
 import "package:photos/ui/common/loading_widget.dart";
+import "package:photos/ui/notification/toast.dart";
 import "package:photos/ui/viewer/file/video_widget_media_kit_common.dart"
     as common;
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/file_util.dart";
-import "package:photos/utils/toast_util.dart";
 
 class VideoWidgetMediaKitNew extends StatefulWidget {
   final EnteFile file;
