+++ mobile/lib/ui/viewer/file_details/upload_icon_widget.dart
@@ -16,7 +16,7 @@ import "package:photos/services/ignored_files_service.dart";
 import "package:photos/services/sync/remote_sync_service.dart";
 import "package:photos/services/sync/sync_service.dart";
 import "package:photos/ui/common/loading_widget.dart";
-import "package:photos/utils/toast_util.dart";
+import "package:photos/ui/notification/toast.dart";
 
 class UploadIconWidget extends StatefulWidget {
   final EnteFile file;
