+++ mobile/lib/ui/viewer/people/save_or_edit_person.dart
@@ -29,6 +29,7 @@ import "package:photos/ui/common/loading_widget.dart";
 import "package:photos/ui/components/action_sheet_widget.dart";
 import "package:photos/ui/components/buttons/button_widget.dart";
 import "package:photos/ui/components/models/button_type.dart";
+import "package:photos/ui/notification/toast.dart";
 import "package:photos/ui/sharing/album_share_info_widget.dart";
 import "package:photos/ui/sharing/user_avator_widget.dart";
 import "package:photos/ui/viewer/file/no_thumbnail_widget.dart";
@@ -40,7 +41,6 @@ import "package:photos/ui/viewer/search/result/person_face_widget.dart";
 import "package:photos/utils/dialog_util.dart";
 import "package:photos/utils/navigation_util.dart";
 import "package:photos/utils/person_contact_linking_util.dart";
-import "package:photos/utils/toast_util.dart";
 
 class SaveOrEditPerson extends StatefulWidget {
   final String? clusterID;
